# frontend-test-go-server/api/swagger.yaml
...
swagger: "2.0"
info:
  description: "The registry API for SwaggerHub"
  version: "1.0.97"
  title: "SwaggerHub Registry API"
  contact:
    name: "SwaggerHub"
    url: "http://swaggerhub.com"
    email: "info@swaggerhub.com"
host: "dev-api.swaggerhub.com"
...
schemes:
- "https"
produces:
- "application/json"
paths:
...
  /plugins:
    get:
      tags:
      - "Plugins"
      summary: "Retrieves a list of all available plugins (ignore system plugins)"
      description: ""
      operationId: "getAvailablePlugins"
      produces:
      - "application/json"
      parameters:
      - name: "owner"
        in: "query"
        required: false
        type: "string"
        x-exportParamName: "Owner"
        x-optionalDataType: "String"
      responses:
        "200":
          description: "a list of plugins in json format"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/PluginDefinition"
        "403":
          description: "access denied"
      security:
      - TokenSecured: []
      x-inflector-hidden: true
      x-swagger-router-controller: "com.smartbear.swaggerhub.registry.service.PluginsV1Service"
    post:
      tags:
      - "Plugins"
      summary: "Register a plugin"
      description: ""
      operationId: "registerPlugin"
      consumes:
      - "application/json"
      - "application/yaml"
      parameters:
      - in: "body"
        name: "pluginDefinition"
        description: "the Plugin definition to be registered or disabled"
        required: true
        schema:
          $ref: "#/definitions/PluginDefinition"
        x-exportParamName: "PluginDefinition"
      responses:
        "200":
          description: "the Plugin was successfully registered"
        "400":
          description: "the Plugin definition was invalid"
        "403":
          description: "access denied"
        "409":
          description: "cannot add plugin with same class and version of existing"
        "415":
          description: "invalid content type"
      security:
      - TokenSecured: []
      x-swagger-router-controller: "com.smartbear.swaggerhub.registry.service.PluginsV1Service"
      x-inflector-hidden: true
  /plugins/{pluginId}:
    delete:
      tags:
      - "Plugins"
      summary: "Disables a plugin"
      description: ""
      operationId: "disablePlugin"
      consumes:
      - "application/json"
      - "application/yaml"
      parameters:
      - name: "pluginId"
        in: "path"
        required: true
        type: "string"
        x-exportParamName: "PluginId"
      responses:
        "200":
          description: "the Plugin was successfully disabled"
        "403":
          description: "access denied"
        "404":
          description: "the Plugin definition was not found"
      security:
      - TokenSecured: []
      x-swagger-router-controller: "com.smartbear.swaggerhub.registry.service.PluginsV1Service"
      x-inflector-hidden: true
  /plugins/configurations:
    get:
      tags:
      - "Plugins"
      summary: "Retrieves a list of enabled plugin configurations"
      description: "Retrieves a list of enabled plugin configurations"
      operationId: "getPlugins"
      produces:
      - "application/json"
      parameters:
      - name: "objectId"
        in: "query"
        description: "plugin configuration objectId"
        required: true
        type: "string"
        x-exportParamName: "ObjectId"
      - name: "page"
        in: "query"
        description: "page to return"
        required: false
        type: "integer"
        default: 0
        format: "int32"
        x-exportParamName: "Page"
        x-optionalDataType: "Int32"
      - name: "limit"
        in: "query"
        description: "number of results per page (1 .. 100)"
        required: false
        type: "integer"
        default: 10
        format: "int32"
        x-exportParamName: "Limit"
        x-optionalDataType: "Int32"
      responses:
        "200":
          description: "a list of plugin configurations in json format"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/PluginConfiguration"
        "403":
          description: "access denied"
      security:
      - TokenSecured: []
      x-inflector-hidden: true
      x-swagger-router-controller: "com.smartbear.swaggerhub.registry.service.PluginsV1Service"
    post:
      tags:
      - "Plugins"
      summary: "Saves the provided Plugin configuration"
      description: "Saves the provided Plugin configuration"
      operationId: "addPluginConfiguration"
      consumes:
      - "application/json"
      - "application/yaml"
      parameters:
      - in: "body"
        name: "pluginConfiguration"
        description: "the Plugin configuration to be added or updated"
        required: true
        schema:
          $ref: "#/definitions/PluginConfiguration"
        x-exportParamName: "PluginConfiguration"
      - name: "trigger"
        in: "query"
        description: "if true, also execute plugin"
        required: false
        type: "boolean"
        default: false
        x-exportParamName: "Trigger"
        x-optionalDataType: "Bool"
      responses:
        "200":
          description: "the Plugin configuration was successfully saved"
        "400":
          description: "the Plugin configuration was invalid"
        "403":
          description: "access denied"
        "409":
          description: "cannot add plugin configuration with same id"
        "415":
          description: "invalid content type"
      security:
      - TokenSecured: []
      x-inflector-hidden: true
      x-swagger-router-controller: "com.smartbear.swaggerhub.registry.service.PluginsV1Service"
    put:
      tags:
      - "Plugins"
      summary: "Updated the provided Plugin configuration"
      description: "Updated the provided Plugin configuration"
      operationId: "updatePluginConfiguration"
      consumes:
      - "application/json"
      - "application/yaml"
      parameters:
      - in: "body"
        name: "pluginConfiguration"
        description: "the Plugin configuration to be added or updated"
        required: true
        schema:
          $ref: "#/definitions/PluginConfiguration"
        x-exportParamName: "PluginConfiguration"
      - name: "trigger"
        in: "query"
        description: "if true, also execute plugin"
        required: false
        type: "boolean"
        default: false
        x-exportParamName: "Trigger"
        x-optionalDataType: "Bool"
      responses:
        "200":
          description: "the Plugin configuration was successfully updated"
        "400":
          description: "the Plugin configuration was invalid"
        "403":
          description: "access denied"
        "404":
          description: "the Plugin configuration was not found"
        "415":
          description: "invalid content type"
      security:
      - TokenSecured: []
      x-inflector-hidden: true
      x-swagger-router-controller: "com.smartbear.swaggerhub.registry.service.PluginsV1Service"
  /plugins/configurations/{pluginId}:
    delete:
      tags:
      - "Plugins"
      summary: "Deletes the provided Plugin configuration"
      description: "Deletes the provided Plugin configuration"
      operationId: "removePluginConfiguration"
      parameters:
      - name: "pluginId"
        in: "path"
        description: "Id of plugin configuration to delete"
        required: true
        type: "string"
        x-exportParamName: "PluginId"
      responses:
        "200":
          description: "the Plugin configuration was successfully deleted"
        "403":
          description: "access denied"
        "404":
          description: "the Plugin configuration was not found"
      security:
      - TokenSecured: []
      x-inflector-hidden: true
      x-swagger-router-controller: "com.smartbear.swaggerhub.registry.service.PluginsV1Service"
  /plugins/configurations/{owner}/options:
    put:
      tags:
      - "Plugins"
      - "System"
      summary: "For a given Owner, make changes to configurations based on the options\
        \ given"
      description: "Using this endpoint allows the user to make mass updates to a\
        \ given owners plugin configurations"
      operationId: "acceptConfigurationOptions"
      consumes:
      - "application/json"
      parameters:
      - name: "owner"
        in: "path"
        description: "API or Domain owner identifier (case-sensitive)"
        required: true
        type: "string"
        x-exportParamName: "Owner"
      - in: "body"
        name: "configurationOptions"
        description: "Using this endpoint allows the user to make mass updates to\
          \ a given owners plugin configurations"
        required: false
        schema:
          $ref: "#/definitions/PluginConfigurationUpdateOptions"
        x-exportParamName: "ConfigurationOptions"
      responses:
        "200":
          description: "Options were accepted and update(s) were made succefully"
        "403":
          description: "User is forbidden from accessing this operation"
        "404":
          description: "Not found"
        "500":
          description: "Unexpected Error"
      security:
      - TokenSecured: []
      x-swagger-router-controller: "com.smartbear.swaggerhub.registry.service.PluginsV1Service"
      x-inflector-hidden: true
  /plugins/configurations/{owner}/{api}/{version}/{integrationPluginType}:
    put:
      tags:
      - "Plugins"
      summary: "Creates or updates a plugin configuration tailored to defined integrationPluginType"
      description: "Creates or updates a plugin configuration tailored to defined\
        \ integrationPluginType"
      operationId: "saveIntegrationPluginOfType"
      consumes:
      - "application/json"
      parameters:
      - name: "owner"
        in: "path"
        description: "API or Domain owner identifier (case-sensitive)"
        required: true
        type: "string"
        x-exportParamName: "Owner"
      - name: "api"
        in: "path"
        description: "API identifier"
        required: true
        type: "string"
        x-exportParamName: "Api"
      - name: "version"
        in: "path"
        description: "version identifier"
        required: true
        type: "string"
        x-exportParamName: "Version"
      - name: "integrationPluginType"
        in: "path"
        description: "What type of plugin to be created for the API"
        required: true
        type: "string"
        enum:
        - "GITHUB"
        - "BITBUCKET"
        - "AZURE_DEVOPS_SERVICES"
        - "AZURE_DEVOPS_SERVER"
        - "GITLAB"
        x-exportParamName: "IntegrationPluginType"
      - name: "oas"
        in: "query"
        description: "The OpenApi Specification (oas) version.\n* 2.0 - Formally known\
          \ as 'swagger'\n* 3.0.x - The latest supported OpenApi Specification version\
          \ by SwaggerHub.\nFor search queries all 3.0.x versions are grouped together\n"
        required: false
        type: "string"
        enum:
        - "2.0"
        - "3.0.0"
        - "3.0.1"
        - "3.0.2"
        - "3.0.3"
        x-exportParamName: "Oas"
        x-optionalDataType: "String"
      - in: "body"
        name: "integrationPluginTypeConfiguration"
        required: true
        schema:
          $ref: "#/definitions/IntegrationPluginConfiguration"
        x-exportParamName: "IntegrationPluginTypeConfiguration"
      responses:
        "200":
          description: "the Plugin configuration was successfully created"
        "400":
          description: "the Plugin configuration was invalid"
        "403":
          description: "access denied"
        "404":
          description: "owner, api, version combination not found"
        "415":
          description: "invalid content type"
      security:
      - TokenSecured: []
      deprecated: true
      x-swagger-router-controller: "com.smartbear.swaggerhub.registry.service.PluginsV1Service"
      x-inflector-hidden: true
  /plugins/configurations/{owner}/{api}/{version}/copy:
    post:
      tags:
      - "Plugins"
      summary: "Options for configurations for a given owner, api, version"
      description: ""
      operationId: "copyConfigurations"
      consumes:
      - "application/json"
      - "application/yaml"
      parameters:
      - name: "owner"
        in: "path"
        description: "API or Domain owner identifier (case-sensitive)"
        required: true
        type: "string"
        x-exportParamName: "Owner"
      - name: "api"
        in: "path"
        description: "API identifier"
        required: true
        type: "string"
        x-exportParamName: "Api"
      - name: "version"
        in: "path"
        description: "version identifier"
        required: true
        type: "string"
        x-exportParamName: "Version"
      - in: "body"
        name: "configurationIds"
        required: false
        schema:
          $ref: "#/definitions/CopyConfigurationDetails"
        x-exportParamName: "ConfigurationIds"
      responses:
        "200":
          description: "the Plugin configuration draft successfully saved"
        "400":
          description: "the request body was invalid"
        "403":
          description: "access denied"
        "415":
          description: "invalid content type"
      security:
      - TokenSecured: []
      x-swagger-router-controller: "com.smartbear.swaggerhub.registry.service.PluginsV1Service"
      x-inflector-hidden: true
  /plugins/configurations/draft:
    post:
      tags:
      - "Plugins"
      summary: "Saves the provided Plugin configuration draft"
      description: "Saves the provided Plugin configuration draft"
      operationId: "addPluginConfigurationDraft"
      consumes:
      - "application/json"
      - "application/yaml"
      parameters:
      - in: "body"
        name: "pluginConfiguration"
        description: "the Plugin configuration to be added or updated"
        required: true
        schema:
          $ref: "#/definitions/PluginConfiguration"
        x-exportParamName: "PluginConfiguration"
      responses:
        "200":
          description: "the Plugin configuration draft successfully saved"
          schema:
            $ref: "#/definitions/PluginConfiguration"
        "400":
          description: "the Plugin configuration was invalid"
        "403":
          description: "access denied"
        "409":
          description: "cannot add plugin configuration with same id"
        "415":
          description: "invalid content type"
      security:
      - TokenSecured: []
      x-swagger-router-controller: "com.smartbear.swaggerhub.registry.service.PluginsV1Service"
      x-inflector-hidden: true
  /plugins/configurations/schema:
    post:
      tags:
      - "Plugins"
      summary: "get configuration schema for the provided Plugin configuration"
      description: "Get configuration schema for the provided Plugin configuration"
      operationId: "buildConfigurationSchema"
      consumes:
      - "application/json"
      - "application/yaml"
      parameters:
      - in: "body"
        name: "pluginConfiguration"
        description: "the Plugin configuration to be added or updated"
        required: true
        schema:
          $ref: "#/definitions/PluginConfiguration"
        x-exportParamName: "PluginConfiguration"
      responses:
        "200":
          description: "the configuration schema in json format"
          schema:
            type: "object"
        "400":
          description: "the Plugin configuration was invalid"
        "403":
          description: "access denied"
        "415":
          description: "invalid content type"
      security:
      - TokenSecured: []
      x-inflector-hidden: true
      x-swagger-router-controller: "com.smartbear.swaggerhub.registry.service.PluginsV1Service"
  /plugins/configurations/item/{id}:
    get:
      tags:
      - "Plugins"
      summary: "returns plugin configuration identified by id"
      description: "returns plugin configuration identified by id"
      operationId: "loadPluginConfiguration"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "plugin configuration id"
        required: true
        type: "string"
        x-exportParamName: "Id"
      responses:
        "200":
          description: "the Plugin configuration identified by id"
          schema:
            $ref: "#/definitions/PluginConfiguration"
        "403":
          description: "access denied"
      security:
      - TokenSecured: []
      x-swagger-router-controller: "com.smartbear.swaggerhub.registry.service.PluginsV1Service"
      x-inflector-hidden: true
  /plugins/configurations/execute/{id}:
    post:
      tags:
      - "Plugins"
      summary: "triggers execution of plugin configuration identified by id"
      description: "triggers execution of plugin configuration identified by id"
      operationId: "triggerPluginConfiguration"
      produces:
      - "application/json"
      parameters:
      - name: "id"
        in: "path"
        description: "plugin configuration id"
        required: true
        type: "string"
        x-exportParamName: "Id"
      responses:
        "200":
          description: "successful execution"
        "403":
          description: "access denied"
      security:
      - TokenSecured: []
      x-inflector-hidden: true
      x-swagger-router-controller: "com.smartbear.swaggerhub.registry.service.PluginsV1Service"
  /plugins/configurations/oauth/{userUuid}/{username}/{pluginConfigId}/{oauthProviderId}:
    get:
      tags:
      - "Plugins"
      summary: "oauth callback url for plugins"
      description: "oauth callback url for plugins"
      operationId: "oauthCallback"
      parameters:
      - name: "userUuid"
        in: "path"
        description: "user uuid originally issueing the request"
        required: true
        type: "string"
        x-exportParamName: "UserUuid"
      - name: "username"
        in: "path"
        description: "user originally issueing the request"
        required: true
        type: "string"
        x-exportParamName: "Username"
      - name: "pluginConfigId"
        in: "path"
        description: "plugin configuration temporary id"
        required: true
        type: "string"
        x-exportParamName: "PluginConfigId"
      - name: "oauthProviderId"
        in: "path"
        description: "oAuth provider id (e.g. github)"
        required: true
        type: "string"
        x-exportParamName: "OauthProviderId"
      - name: "code"
        in: "query"
        description: "oAuth callback code param"
        required: false
        type: "string"
        x-exportParamName: "Code"
        x-optionalDataType: "String"
      - name: "state"
        in: "query"
        description: "oAuth callback state param"
        required: false
        type: "string"
        x-exportParamName: "State"
        x-optionalDataType: "String"
      responses:
        "302":
          description: "302 response"
          headers:
            Location:
              type: "string"
      x-swagger-router-controller: "com.smartbear.swaggerhub.registry.service.PluginsV1Service"
      x-inflector-hidden: true
  /plugins/configurations/oauth:
    get:
      tags:
      - "Plugins"
      summary: "oauth callback url for plugins with query params"
      description: "oauth callback url for plugins with query params"
      operationId: "oauthCallbackWithQueryParams"
      parameters:
      - name: "userUuid"
        in: "query"
        description: "user uuid originally issueing the request"
        required: true
        type: "string"
        x-exportParamName: "UserUuid"
      - name: "username"
        in: "query"
        description: "user originally issuing the request"
        required: true
        type: "string"
        x-exportParamName: "Username"
      - name: "pluginConfigId"
        in: "query"
        description: "plugin configuration temporary id"
        required: true
        type: "string"
        x-exportParamName: "PluginConfigId"
      - name: "oauthProviderId"
        in: "query"
        description: "oAuth provider id (e.g. github)"
        required: true
        type: "string"
        x-exportParamName: "OauthProviderId"
      - name: "code"
        in: "query"
        description: "oAuth callback code param"
        required: false
        type: "string"
        x-exportParamName: "Code"
        x-optionalDataType: "String"
      - name: "state"
        in: "query"
        description: "oAuth callback state param"
        required: false
        type: "string"
        x-exportParamName: "State"
        x-optionalDataType: "String"
      responses:
        "302":
          description: "302 response"
          headers:
            Location:
              type: "string"
      x-swagger-router-controller: "com.smartbear.swaggerhub.registry.service.PluginsV1Service"
      x-inflector-hidden: true
  /plugins/configurations/oauthWithState:
    get:
      tags:
      - "Plugins"
      summary: "oauth callback url for plugins with a state query parameter"
      description: "oauth callback url for plugins with a state query parameter"
      operationId: "oauthCallbackWithStateOnly"
      parameters:
      - name: "code"
        in: "query"
        description: "oAuth callback code param"
        required: false
        type: "string"
        pattern: "[^|]+(\\|[^|]+){4}"
        x-exportParamName: "Code"
        x-optionalDataType: "String"
      - name: "state"
        in: "query"
        description: "oAuth callback state parameter in the following form `state|userId|userName|pluginConfigurationId|oauthProviderId`"
        required: false
        type: "string"
        x-exportParamName: "State"
        x-optionalDataType: "String"
      responses:
        "302":
          description: "302 response"
          headers:
            Location:
              type: "string"
      x-swagger-router-controller: "com.smartbear.swaggerhub.registry.service.PluginsV1Service"
      x-inflector-hidden: true
  /plugins/systemconfigurations:
    get:
      tags:
      - "Plugins"
      summary: "Gets system configurations"
      operationId: "getSystemPluginConfigurations"
      consumes:
      - "application/json"
      - "application/yaml"
      parameters: []
      responses:
        "200":
          description: "System Plugin configuration was successfully retrieved"
          schema:
            type: "array"
            items:
              $ref: "#/definitions/SystemPluginConfiguration"
        "403":
          description: "access denied"
      security:
      - TokenSecured: []
      x-swagger-router-controller: "com.smartbear.swaggerhub.registry.service.PluginsV1Service"
      x-inflector-hidden: true
    post:
      tags:
      - "Plugins"
      summary: "Saves the provided System Plugin configuration"
      description: "Saves the provided System Plugin configuration"
      operationId: "addSystemPluginConfiguration"
      consumes:
      - "application/json"
      - "application/yaml"
      parameters:
      - in: "body"
        name: "systemPluginConfiguration"
        description: "the System Plugin configuration to be added or updated"
        required: true
        schema:
          $ref: "#/definitions/SystemPluginConfiguration"
        x-exportParamName: "SystemPluginConfiguration"
      responses:
        "200":
          description: "the System Plugin configuration was successfully saved"
        "400":
          description: "the System Plugin configuration was invalid"
        "403":
          description: "access denied"
        "409":
          description: "cannot add System Plugin configuration with same id"
        "415":
          description: "invalid content type"
      security:
      - TokenSecured: []
      x-swagger-router-controller: "com.smartbear.swaggerhub.registry.service.PluginsV1Service"
      x-inflector-hidden: true
    put:
      tags:
      - "Plugins"
      summary: "Updates the provided System Plugin configuration"
      description: "Updated the provided System Plugin configuration"
      operationId: "updateSystemPluginConfiguration"
      consumes:
      - "application/json"
      - "application/yaml"
      parameters:
      - in: "body"
        name: "systemPluginConfiguration"
        description: "the System Plugin configuration to be added or updated"
        required: true
        schema:
          $ref: "#/definitions/SystemPluginConfiguration"
        x-exportParamName: "SystemPluginConfiguration"
      responses:
        "200":
          description: "the System Plugin configuration was successfully updated"
        "400":
          description: "the System Plugin configuration was invalid"
        "403":
          description: "access denied"
        "404":
          description: "the System Plugin configuration was not found"
        "415":
          description: "invalid content type"
      security:
      - TokenSecured: []
      x-swagger-router-controller: "com.smartbear.swaggerhub.registry.service.PluginsV1Service"
      x-inflector-hidden: true
...
securityDefinitions:
  TokenSecured:
    description: "Specify your SwaggerHub API key found in https://app.swaggerhub.com/settings/apiKey."
    type: "apiKey"
    name: "Authorization"
    in: "header"
definitions:
...
  PluginDefinition:
    type: "object"
    required:
    - "configurationSchema"
    - "createdBy"
    - "createdOn"
    - "enabled"
    - "implementingClass"
    - "name"
    - "version"
    properties:
      id:
        type: "string"
      name:
        type: "string"
      implementingClass:
        type: "string"
      version:
        type: "string"
      createdBy:
        type: "string"
      createdOn:
        type: "string"
        format: "date-time"
      enabled:
        type: "boolean"
      configurationSchema:
        type: "object"
        properties: {}
      limit:
        type: "integer"
        format: "int64"
        description: "count of allowed configurations"
    example:
      createdBy: "createdBy"
      name: "name"
      configurationSchema: "{}"
      limit: 0
      id: "id"
      version: "version"
      createdOn: "2000-01-23T04:56:07.000+00:00"
      enabled: true
      implementingClass: "implementingClass"
    x-swagger-router-model: "com.smartbear.swaggerhub.plugins.model.PluginDefinition"
  PluginConfiguration:
    type: "object"
    required:
    - "definitionId"
    - "lifecycles"
    - "objectId"
    - "ownerName"
    - "path"
    properties:
      id:
        type: "string"
        format: "UUID"
      name:
        type: "string"
      objectId:
        type: "string"
      lifecycles:
        type: "array"
        items:
          type: "string"
      configuration:
        type: "object"
        properties: {}
      ownerName:
        type: "string"
      path:
        type: "string"
      definitionId:
        type: "string"
      draft:
        type: "boolean"
    example:
      path: "path"
      ownerName: "ownerName"
      configuration: "{}"
      draft: true
      name: "name"
      id: "id"
      lifecycles:
      - "lifecycles"
      - "lifecycles"
      objectId: "objectId"
      definitionId: "definitionId"
    x-swagger-router-model: "com.smartbear.swaggerhub.plugins.model.PluginConfiguration"
  SystemPluginConfiguration:
    type: "object"
    required:
    - "implementingClass"
    - "lifecycles"
    - "version"
    properties:
      id:
        type: "string"
        format: "UUID"
      lifecycles:
        type: "array"
        items:
          type: "string"
      configuration:
        type: "object"
        additionalProperties:
          type: "string"
      implementingClass:
        type: "string"
      version:
        type: "string"
    example:
      configuration:
        key: "configuration"
      id: "id"
      lifecycles:
      - "lifecycles"
      - "lifecycles"
      version: "version"
      implementingClass: "implementingClass"
    x-swagger-router-model: "com.smartbear.swaggerhub.plugins.model.SystemPluginConfiguration"
...
  IntegrationPluginConfiguration:
    type: "object"
    required:
    - "configType"
    discriminator: "configType"
    properties:
      configType:
        type: "string"
    description: "Can be used as the parent superclass for any plugin integrations\
      \ which are available for usage as part of SwaggerHub."
    example:
      configType: "configType"
...
  PluginConfigurationUpdateOptions:
    type: "object"
    properties:
      disable:
        type: "string"
        description: "Available options\nPLAN = disable plugins based on the current\
          \ owner's plan. See plan details for more info on specific plugins allowed\
          \ per plan\n"
        enum:
        - "PLAN"
    example:
      disable: "PLAN"
  CopyConfigurationDetails:
    type: "object"
    properties:
      version:
        type: "string"
        example: "1.0.1"
        description: "The version of the API to copy the configurations to."
      integrationIds:
        type: "array"
        items:
          type: "string"
    example:
      integrationIds:
      - "integrationIds"
      - "integrationIds"
      version: "1.0.1"
...