# frontend-test-go-server/go/model_copy_configuration_details.go
/*
 * SwaggerHub Registry API
 *
 * The registry API for SwaggerHub
 *
 * API version: 1.0.97
 * Contact: info@swaggerhub.com
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type CopyConfigurationDetails struct {

	// The version of the API to copy the configurations to.
	Version string `json:"version,omitempty"`

	IntegrationIds []string `json:"integrationIds,omitempty"`
}