# frontend-test-go-server/go/model_system_plugin_configuration.go
/*
 * SwaggerHub Registry API
 *
 * The registry API for SwaggerHub
 *
 * API version: 1.0.97
 * Contact: info@swaggerhub.com
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */

package swagger

type SystemPluginConfiguration struct {

	Id string `json:"id,omitempty"`

	Lifecycles []string `json:"lifecycles"`

	Configuration map[string]string `json:"configuration,omitempty"`

	ImplementingClass string `json:"implementingClass"`

	Version string `json:"version"`
}